requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.118.0
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
# Dependencias para base de datos
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
sqlalchemy-cockroachdb==2.0.3
sqlalchemy==1.4.48
psycopg2-binary==2.9.7